requirements.txt
torch==2.0.1
torchvision==0.15.2
pillow==9.2.0
numpy==1.23.5
scikit-learn==1.1.3
tqdm==4.64.1
